feat: support named pipes
Currently the `getPort` method is always called if `port` is set to anything but `0`. This causes an issue because `getPort` calls `const port = server.address().port` which doesn't return the correct value when using named pipes. This is especially important for use cases involving `iisnode` and similar that interact with the node process via a named pipe as the port instead.